+++ src/BeastyBotMain.py
@@ -2,8 +2,8 @@ import subprocess
 from twitchio.ext import commands
 
 from src.helperfunc.base_values import CHANNEL_NAME, setup_logger, get_valid_token
-from helperfunc.object_manager import UserManager, ObjectManager
-from helperfunc.setting import BotList
+from src.helperfunc.object_manager import UserManager, ObjectManager
+from src.helperfunc.setting import BotList
 
 # TODO: Channelpoint integration
