+++ src/dustbunny_gen_bot.py
@@ -4,7 +4,7 @@ from twitchio.ext import commands
 import os, sys
 from src.helperfunc.global_methods import calculate_interest, do_the_cleaning_command, generate_rnd_amount_to_steal
 from src.helperfunc.base_values import CHANNEL_NAME, setup_logger, get_valid_token
-from helperfunc.object_manager import ObjectManager, UserManager
+from src.helperfunc.object_manager import ObjectManager, UserManager
 
 _logger = setup_logger(__name__)
